# Speed up decoding of bytes/string tuple elements.