--- a/foundationdb-haskell.cabal
+++ b/foundationdb-haskell.cabal
@@ -161,4 +161,5 @@ benchmark tuple-bench
                        , foundationdb-haskell
                        , gauge
                        , bytestring >= 0.10.8.2 && <0.11
+                       , text
   default-language:    Haskell2010